requirements.txt
beautifulsoup4==4.12.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
packaging==24.2
pillow==11.2.1
pytesseract==0.3.13
requests==2.31.0
setuptools==78.1.0
soupsieve==2.6
urllib3==2.4.0
wheel==0.45.1